+++ soat-immo-web/src/test/java/fr/soat/interco/web/test/service/BienImmoServiceTest.java
@@ -111,7 +111,7 @@ public class BienImmoServiceTest extends ParentTests {
         Assertions.assertThat(immoByType).hasSize(NB_BIEN_IMMO);
 
         //test delete bienImmo
-        Integer idbien_immobiler = bienImmobilier.getIdBien();
+        Long idbien_immobiler = bienImmobilier.getIdBien();
         bienImmoService.deleteBienImmo(bienImmobilier);
         BienImmobilier bienImmoById = bienImmoService.findBienImmoById(idbien_immobiler);
         Assertions.assertThat(bienImmoById).isNull();
